requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
attrs==23.1.0
Automat==22.10.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
debugpy @ file:///D:/bld/debugpy_1680755654494/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dnspython==2.3.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock==3.12.1
hyperlink==21.0.0
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
incremental==22.10.0
ipykernel @ file:///D:/bld/ipykernel_1655241698613/work
ipython @ file:///D:/bld/ipython_1653755095589/work
itemadapter==0.8.0
itemloaders==1.1.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jmespath==1.0.1
joblib==1.2.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///D:/bld/jupyter_core_1678994400417/work
lxml==4.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.24.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.2
parsel==1.8.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1686403120743/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
Protego==0.2.1
psutil @ file:///D:/bld/psutil_1681775154843/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pymongo==4.3.3
pyOpenSSL==23.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz==2023.3
pywin32==304
pyzmq @ file:///D:/bld/pyzmq_1685519522122/work
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.9.0
service-identity==21.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
threadpoolctl==3.1.0
tldextract==3.4.4
tornado @ file:///D:/bld/tornado_1684150254720/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
tzdata==2023.3
urllib3==2.0.3
w3lib==2.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
xgboost==1.7.5
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
zope.interface==6.0
